Module eant2::options

Option types for the EANT2 algorithm.

Structs

EANT2Options
A container for all parameters and options for the EANT2 algorithm. Using default options is convenient, but may slightly reduce the quality of the neural networks generated, and can increase the time needed to find a solution. It is recommended to adjust the parameters depending on whether it is more important to find a solution quickly or to find a better solution. Note that increasing population size and CMA-ES runs will only affect the quality a small amount, so using smaller values will greatly speed up the algorithm, and does not have a large downside.