Struct eant2::options::EANT2Options

pub struct EANT2Options {
    pub inputs: usize,
    pub outputs: usize,
    pub population_size: usize,
    pub offspring_count: usize,
    pub min_fitness: f64,
    pub max_generations: usize,
    pub threads: u8,
    pub cmaes_runs: usize,
    pub cmaes_end_conditions: Vec<CMAESEndConditions>,
}

A container for all parameters and options for the EANT2 algorithm. Using default options is convenient, but may slightly reduce the quality of the neural networks generated, and can increase the time needed to find a solution. It is recommended to adjust the parameters depending on whether it is more important to find a solution quickly or to find a better solution. Note that increasing population size and CMA-ES runs will only affect the quality a small amount, so using smaller values will greatly speed up the algorithm, and does not have a large downside.

Examples

use eant2::*;

// A set of default options, with 2 inputs and 3 outputs to each neural network, and a minimum
// fitness of 10.0
let options = EANT2Options::new(2, 3, 10.0);

// A set of options with 2 CMA-ES optimizations per individual, and a population size of 50
let options = EANT2Options::new(2, 3, 10.0)
    .cmaes_runs(2)
    .population_size(50);

Fields

inputs

outputs

population_size

offspring_count

min_fitness

max_generations

threads

cmaes_runs

cmaes_end_conditions

Methods

impl EANT2Options

fn new(inputs: usize, outputs: usize, min_fitness: f64) -> EANT2Options

Returns a set of default options.

fn population_size(self, size: usize) -> EANT2Options

Sets the population size

fn offspring_count(self, count: usize) -> EANT2Options

Sets the offspring count (how many offspring each individual spawns)

fn min_fitness(self, fitness: f64) -> EANT2Options

Sets the minimum fitness. The algorithm terminates if the best individual has a fitness of at least the specified amount.

fn max_generations(self, generations: usize) -> EANT2Options

Sets the maximum generations. The algorithm terminates after the specified number of generations pass.

fn threads(self, threads: u8) -> EANT2Options

Sets the number of threads to use in the algorithm.

fn cmaes_runs(self, runs: usize) -> EANT2Options

Sets the number of times CMA-ES should be run on each individual. More runs will produce higher quality neural networks faster at the cost of a higher number of fitness function calls. Increase the number of runs if the fitness function is cheap to run, or if time taken to find a solution is not important.

fn cmaes_end_conditions(self, conditions: Vec<CMAESEndConditions>) -> EANT2Options

Sets the end conditions of CMA-ES. Setting higher stable generations or max generations or evaluations can produce higher quality neural networks at the cost of a longer time taken to find a solution. Decreasing these options is a good idea when it is important to find a solution quickly. DO NOT set the fitness threshold option, or it will take a very long time to find a solution, even on the simplest problems! It is recommended to leave this option alone, but the documentation for it is available here.